Rename env.js to envelope.js, and update entries in Gruntfile and app.js
