+++ src/envelope.js
